Backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
pytest>=7.0
pytest-asyncio>=0.21
httpx>=0.27
